refactor(cl): convert SearchAlertsAssertions and V4SearchAPIAssertions to mixins
