+++ cl/opinion_page/tests.py
@@ -119,9 +119,9 @@ class SimpleLoadTest(TestCase):
 
 
 class OpinionPageLoadTest(
-    CourtMixin,
-    PeopleMixin,
     SearchMixin,
+    PeopleMixin,
+    CourtMixin,
     ESIndexTestCase,
     TestCase,
 ):
